+++ HISTORY.rst
@@ -6,6 +6,8 @@ History
 
 * Add fixer for Django 3.1 removal of ``Signal``\’s argument ``providing_args``.
 
+* Add fixer for Django 3.1 requirement to pass ``get_random_string()`` the ``length`` argument.
+
 * Fix Python 3.8 compatibility.
 
 * Drop Python 3.6 and 3.7 support, since they never worked, and the incompatibilities in the ``ast`` module are hard to cover.